Bugs Fixed 02/23/2022
•USERS•
Roles - kapag dinelete yung role, nagiging system administrator yung role ng user.
	Re: Fixed

Dropdown list ng status - kapag nilagay na inactive at sinave, "active" parin ang nakalagay pero yung user ay inactive (di makapag login).
	Re: Fixed

Profile picture - hindi mapalitan yung current profile picture (both admin and user nag error).
	Re: Fixed

•CONTENT MANAGEMENT•
Announcements, Slider, and News and Events - sa pag eedit kailangan may ma-upload na picture.
	Re: Fixed
		Announcements:
 	Can now create and edit announcements without image
Image uploaded can now be viewed in edit
Slider:
	Image is required in sliders
Image uploaded can now be viewed in edit
News and Events:
	Image is required in News and Events
	Image uploaded can now be viewed in edit

•ELECTION MANAGEMENT
Candidates - i-default agad ang sa election galing sa pinili ng user.
	Re: di na kailangang baguhin, matagal lang talaga ng mag loading

- kapag lagpas tatlo na, walang nalabas na error.
Re: May error naman message naman na lumalabas

- Bug! Nawawala yung candidates kapag nag dagdag ng bagong position.
	Re: Fixed

##Voting - sa select management, kapag mag seselect kailangan pa irefresh ang page para makita yung results.
Re: di na kailangang baguhin, matagal lang talaga ng mag loading

Election - sa election details, iniistore parin yung nawalang candidates na nawala kanina sa pag dagdag ng bagong position sa candidates.
	Re: Fixed. Yung error nakakabit sa naunang bug na naayos na

•PAYMENT MANAGEMENT•
Contributions - hindi makapag delete.
	Re: Fixed

•DISCUSSIONS•
- pwede mag add kahit walang subject.
	Re: Fixed
- hindi makapag delete.
Re: Fixed

•REPORTS•
Login reports - kailangan irefresh para makita yung results.
Re: di na kailangang baguhin, matagal lang talaga ng mag loading

Payment reports - error sa generate pdf.
	Re: Fixed

Other Fixes:
	Fixed the format of Generate Contribution PDF;
	Fixed the date range picker of payment pdf generator;

+++ modules/Roles/Controllers/Roles.php
@@ -3,6 +3,7 @@ namespace Modules\Roles\Controllers;
 
 use App\Controllers\BaseController;
 use App\Controllers\Roles as Control;
+use App\Models as RealModels;
 use Modules\Roles\Models as Models;
 use Modules\Permissions\Models as Perms;
 use App\Models as AppModels;
@@ -11,6 +12,7 @@ class Roles extends BaseController
 {
     public function __construct() {
         $this->roleModel = new Models\RoleModel();
+        $this->userModel = new RealModels\UserModel;
         $this->rolePermissionModel = new Perms\RolePermissionModel();
         $this->activityLogModel = new AppModels\ActivityLogModel();
     }
@@ -96,13 +98,18 @@ class Roles extends BaseController
             return redirect()->to(base_url());
         }
         $data['rolePermission'] = $data['perm_id']['rolePermission'];
-        if($this->roleModel->where('id', $id)->delete()) {
-            $activityLog['user'] = $this->session->get('user_id');
-            $activityLog['description'] = 'Deleted a role';
-            $this->activityLogModel->save($activityLog);
-          $this->session->setFlashData('successMsg', 'Successfully deleted role');
+
+        if($this->userModel->where('role', $id)->countAllResults(false) == 0){
+            if($this->roleModel->where('id', $id)->delete()) {
+                $activityLog['user'] = $this->session->get('user_id');
+                $activityLog['description'] = 'Deleted a role';
+                $this->activityLogModel->save($activityLog);
+              $this->session->setFlashData('successMsg', 'Successfully deleted role');
+            } else {
+              $this->session->setFlashData('failMsg', 'Something went wrong!');
+            }
         } else {
-          $this->session->setFlashData('errorMsg', 'Something went wrong!');
+          $this->session->setFlashData('failMsg', 'Cannot delete - there are accounts with this role!');
         }
         return redirect()->to(base_url('admin/roles'));
     }
